Avanzando

+++ Json/ManejoJson.cs
@@ -1,3 +1,4 @@
+using System.Reflection;
 using System.Text.Json;
 using Personajes;
 
@@ -8,34 +9,33 @@ namespace ManejoJson
 
         public static bool Existe(string nombreArchivo)
         {
-            if(File.Exists(nombreArchivo))
+            if (File.Exists(nombreArchivo))
             {
                 return true;
-            }else{
+            }
+            else
+            {
                 return false;
             }
         }
         public static void GuardarPersonajes(List<Personaje> misPersonajes, string nombreArchivo)
         {
             // Creo el directorio si no existe
-            string directorio = Path.GetDirectoryName(nombreArchivo);
-            if (!Directory.Exists(directorio))
+            if (!File.Exists(nombreArchivo))
             {
-                Directory.CreateDirectory(directorio);
+                var opciones = new JsonSerializerOptions
+                {
+                    Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping, //Para evitar los acentos
+                    WriteIndented = true
+                };
+
+                // Convierto la lista de personajes a JSON
+                string jsonString = JsonSerializer.Serialize(misPersonajes, opciones);
+
+                // Guardo el JSON en el archivo
+                File.WriteAllText(nombreArchivo, jsonString);
             }
 
-            // Configuración para formatear el JSON
-            var opciones = new JsonSerializerOptions
-            {
-                Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping, //Para evitar los acentos
-                WriteIndented = true
-            };
-
-            // Convierto la lista de personajes a JSON
-            string jsonString = JsonSerializer.Serialize(misPersonajes, opciones);
-
-            // Guardo el JSON en el archivo
-            File.WriteAllText(nombreArchivo, jsonString);
         }
 
         //No uso esta funcion pero la tengo porque debo hacerla
